move the logic of validating the clone's source away from webhooks
