+++ pkg/virt-controller/watch/clone/clone_base.go
@@ -1,8 +1,10 @@
 package clone
 
 import (
+	"context"
 	"errors"
 	"fmt"
+	"strings"
 	"time"
 
 	k8scorev1 "k8s.io/api/core/v1"
@@ -21,6 +23,11 @@ import (
 	"kubevirt.io/client-go/log"
 
 	"kubevirt.io/kubevirt/pkg/storage/snapshot"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	k8sfield "k8s.io/apimachinery/pkg/util/validation/field"
+
+	"kubevirt.io/kubevirt/pkg/network/link"
 )
 
 type Event string
@@ -42,6 +49,8 @@ const (
 	SourceDoesNotExist              Event = "SourceDoesNotExist"
 	SourceWithBackendStorageInvalid Event = "SourceVMWithBackendStorageInvalid"
 	VMVolumeSnapshotsInvalid        Event = "VMVolumeSnapshotsInvalid"
+	virtualMachineKind                    = "VirtualMachine"
+	virtualMachineSnapshotKind            = "VirtualMachineSnapshot"
 )
 
 var (
@@ -411,4 +420,157 @@ func (ctrl *VMCloneController) Execute() bool {
 func (ctrl *VMCloneController) runWorker() {
 	for ctrl.Execute() {
 	}
+}
+
+func ValidateSourceAndTargetKind(vmClone *clone.VirtualMachineClone) []metav1.StatusCause {
+	var causes []metav1.StatusCause = nil
+	sourceField := k8sfield.NewPath("spec")
+
+	supportedSourceTypes := []string{virtualMachineKind, virtualMachineSnapshotKind}
+	supportedTargetTypes := []string{virtualMachineKind}
+
+	if !doesSliceContainStr(supportedSourceTypes, vmClone.Spec.Source.Kind) {
+		causes = []metav1.StatusCause{{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Source kind is not supported",
+			Field:   sourceField.Child("Source").Child("Kind").String(),
+		}}
+	}
+
+	if vmClone.Spec.Target != nil && !doesSliceContainStr(supportedTargetTypes, vmClone.Spec.Target.Kind) {
+		if causes == nil {
+			causes = []metav1.StatusCause{}
+		}
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Target kind is not supported",
+			Field:   sourceField.Child("Target").Child("Kind").String(),
+		})
+	}
+
+	return causes
+}
+
+func ValidateSource(ctx context.Context, client kubecli.KubevirtClient, vmClone *clone.VirtualMachineClone) []metav1.StatusCause {
+	var causes []metav1.StatusCause = nil
+	sourceField := k8sfield.NewPath("spec")
+	source := vmClone.Spec.Source
+
+	if source == nil {
+		causes = []metav1.StatusCause{{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Source cannot be nil",
+			Field:   sourceField.Child("Source").String(),
+		}}
+		return causes
+	}
+	if source.APIGroup == nil || *source.APIGroup == "" {
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Source's APIGroup cannot be empty",
+			Field:   sourceField.Child("Source").Child("APIGroup").String(),
+		})
+	}
+	if source.Kind == "" {
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Source's Kind cannot be empty",
+			Field:   sourceField.Child("Source").String(),
+		})
+	}
+	if source.Name == "" {
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Source's name cannot be empty",
+			Field:   sourceField.Child("Source").Child("Name").String(),
+		})
+	}
+	return causes
+}
+
+func ValidateTarget(vmClone *clone.VirtualMachineClone) []metav1.StatusCause {
+	var causes []metav1.StatusCause
+
+	source := vmClone.Spec.Source
+	target := vmClone.Spec.Target
+
+	if source != nil &&
+		target != nil &&
+		source.Kind == virtualMachineKind &&
+		target.Kind == virtualMachineKind &&
+		target.Name == source.Name {
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: "Target name cannot be equal to source name when both are VirtualMachines",
+			Field:   k8sfield.NewPath("spec").Child("target").Child("name").String(),
+		})
+	}
+
+	return causes
+}
+
+func ValidateNewMacAddresses(vmClone *clone.VirtualMachineClone) []metav1.StatusCause {
+	var causes []metav1.StatusCause
+
+	for ifaceName, ifaceMac := range vmClone.Spec.NewMacAddresses {
+		if ifaceMac != "" {
+			if err := link.ValidateMacAddress(ifaceMac); err != nil {
+				causes = append(causes, metav1.StatusCause{
+					Type:    metav1.CauseTypeFieldValueInvalid,
+					Message: fmt.Sprintf("interface %s has malformed MAC address (%s).", ifaceName, ifaceMac),
+					Field:   k8sfield.NewPath("spec").Child("newMacAddresses").Child(ifaceName).String(),
+				})
+			}
+		}
+	}
+
+	return causes
+}
+
+func ValidateFilters(filters []string, fieldName string) (causes []metav1.StatusCause) {
+	if filters == nil {
+		return nil
+	}
+
+	addCause := func(message string) {
+		causes = append(causes, metav1.StatusCause{
+			Type:    metav1.CauseTypeFieldValueInvalid,
+			Message: message,
+			Field:   fieldName,
+		})
+	}
+	const negationChar = "!"
+	const wildcardChar = "*"
+
+	for _, filter := range filters {
+		if len(filter) == 1 {
+			if filter == negationChar {
+				addCause("a negation character is not a valid filter")
+			}
+			continue
+		}
+
+		const errPattern = "%s filter %s is invalid: cannot contain a %s character (%s); FilterRules: %s"
+
+		if filterWithoutFirstChar := filter[1:]; strings.Contains(filterWithoutFirstChar, negationChar) {
+			addCause(fmt.Sprintf(errPattern, fieldName, filter, "negation", negationChar, "NegationChar can be only used at the beginning of the filter"))
+		}
+
+		if filterWithoutLastChar := filter[:len(filter)-1]; strings.Contains(filterWithoutLastChar, wildcardChar) {
+			addCause(fmt.Sprintf(errPattern, fieldName, filter, "wildcard", wildcardChar, "WildcardChar can be only at the end of the filter"))
+		}
+	}
+
+	return causes
+}
+
+func doesSliceContainStr(slice []string, str string) (isFound bool) {
+	for _, curSliceStr := range slice {
+		if curSliceStr == str {
+			isFound = true
+			break
+		}
+	}
+
+	return isFound
 }
